PD Plugin Definition

Plugins for gsm.app are defined by a named list with information about the plugin. These can be loaded from a yaml file with gsm.app::plugin_Read(). We provide a ready-to-use plugin definition for this PD plugin.

Usage

pluginPD()

Value

A plugin definition for use in gsm.app.

Examples

pluginPD()
#> $meta
#> $meta$Name
#> [1] "Protocol Deviations"
#> 
#> 
#> $shiny
#> $shiny$UI
#> [1] "mod_PD_UI"
#> 
#> $shiny$Server
#> [1] "mod_PD_Server"
#> 
#> 
#> $domains
#> [1] "PD"   "SUBJ"
#>